Update website layout. Refs #6, refs #9, refs #10
Here, a new layout for the website is made. The navigation bar at the top can be used to navigate to a particular team member's section of the page. Each team member's section also has a button leading to their info_nnnn.txt file as well as their personal website (this needs to be added by each team member). Additionally, a dark salmon theme is proposed here.